+++ src/doitoml/utils/path.py
@@ -3,6 +3,8 @@
 from pathlib import Path
 from typing import Optional, Tuple
 
+from doitoml.types import PathOrString
+
 
 def ensure_parents(*paths: Optional[Path]) -> Tuple[Optional[Path], ...]:
     """Clean out some paths and ensure their parents."""
@@ -11,3 +13,14 @@ def ensure_parents(*paths: Optional[Path]) -> Tuple[Optional[Path], ...]:
             continue
         path.parent.mkdir(parents=True, exist_ok=True)
     return paths
+
+
+def normalize_path(path: PathOrString) -> str:
+    """Apply some best-effort, platform-aware path normalization."""
+    as_path = Path(path).resolve()
+    norm = str(as_path)
+    if as_path.drive:  # pragma: no cover
+        norm_bits = str(path).split(":")
+        norm = ":".join([norm_bits[0].lower(), *norm_bits[1:]])
+        norm = norm.replace("\\", "/")
+    return norm